Made a utility file for node-pty
- Moved the asnc queue to the `lib` folder
- Created a new file for the node-pty utility functions
- Removed old functions
